Updated oninvalid and oninput attribute
Changed the oninvalid and oninput attributes along with the pattern attribute to perform input validation. This change enables the default HTML constraint API message to be altered and thus display a customized message instead.
Also changed is the insertion of a footer instructing the user to reach the end of the form to trigger the visibility of the submit button.