reset button, footer and input invalid
This alteration consists of the addition of the CSS responsible for the reset button which has been positioned and also the display of the option button as a block. The input: invalid pseudo-selector has been included to style the input box and change the border with a red box-shadow and thus inform the user visually when an error has been comited upon value input.